Configured start_grpc_server functionality and moved all  CRUD methods into a separate module
